add likesCount and commentsCount to Home & move logged in check to userActions
