+++ src/App.js
@@ -11,40 +11,39 @@ import {
 import SignupForm from './components/SignupForm'
 import LoginForm from './components/LoginForm'
 import ExploreImagesContainer from './containers/ExploreImagesContainer';
-import { setUser } from './actions/userActions'
+import { setUser, setUserIfAlreadyLoggedIn } from './actions/userActions'
 import Navbar from './components/Navbar'
 import Home from './components/Home'
 
 const url = "http://localhost:3001/api/v1/logged_in"
 
-function App({images, loading, fetchImages, user, setUser}) {
-  
+function App({images, loading, fetchImages, user, setUser, setUserIfAlreadyLoggedIn}) {
   // Check to see if the user is still logged in at app mount
   useEffect(() => {
-    setUserIfAlreadyLoggedIn()
+    setUserIfAlreadyLoggedIn();
   }, [])
 
-  const setUserIfAlreadyLoggedIn = () => {
-    const token = localStorage.getItem("token")
-    if (token){
-      fetch(url, {
-        headers: {
-          Authorization: `Bearer ${token}`
-        }
-      })
-      .then(resp => {
-        if (!resp.ok){
-          throw Error(resp.statusText)
-        } else {
-          return resp.json();
-        }
-      })
-      .then(returnUserData => {
-        setUser(returnUserData)
-      })
-      .catch(error => console.log(error))
-    }
-  }
+  // const setUserIfAlreadyLoggedIn = () => {
+  //   const token = localStorage.getItem("token")
+  //   if (token){
+  //     fetch(url, {
+  //       headers: {
+  //         Authorization: `Bearer ${token}`
+  //       }
+  //     })
+  //     .then(resp => {
+  //       if (!resp.ok){
+  //         throw Error(resp.statusText)
+  //       } else {
+  //         return resp.json();
+  //       }
+  //     })
+  //     .then(returnUserData => {
+  //       setUser(returnUserData)
+  //     })
+  //     .catch(error => console.log(error))
+  //   }
+  // }
 
   // useEffect(() => {
   //   fetchImages()
@@ -68,7 +67,7 @@ function App({images, loading, fetchImages, user, setUser}) {
           </Route>
           
           <Route path="/login">
-            <LoginForm />
+          { loggedIn ? <Redirect to="/explore" /> : <LoginForm />}
           </Route>
 
           <Route exact path="/home">
@@ -85,4 +84,4 @@ function App({images, loading, fetchImages, user, setUser}) {
   );
 }
 
-export default connect(({images, loading, user}) => ({images, loading, user}), { fetchImages, setUser })(App);
+export default connect(({images, loading, user}) => ({images, loading, user}), { fetchImages, setUser, setUserIfAlreadyLoggedIn })(App);
